+++ apps/api/src/app/portfolio/portfolio.service.ts
@@ -957,6 +957,18 @@ export class PortfolioService {
       marketPriceMin,
       SymbolProfile,
       tags,
+      assetProfile: {
+        assetClass: SymbolProfile.assetClass,
+        assetSubClass: SymbolProfile.assetSubClass,
+        countries: SymbolProfile.countries,
+        currency: SymbolProfile.currency,
+        dataSource: SymbolProfile.dataSource,
+        isin: SymbolProfile.isin,
+        name: SymbolProfile.name,
+        sectors: SymbolProfile.sectors,
+        symbol: SymbolProfile.symbol,
+        userId: SymbolProfile.userId
+      },
       averagePrice: averagePrice.toNumber(),
       dataProviderInfo: portfolioCalculator.getDataProviderInfos()?.[0],
       dividendInBaseCurrency: dividendInBaseCurrency.toNumber(),